pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "flask-config"
authors = [
    {name = "Linn Aung Htet",email = "mtm.linnaunghtet@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "werkzeug (>=3.1.5,<4.0.0)",
    "flask-cors (>=6.0.2,<7.0.0)",
    "flask-mail (>=0.10.0,<0.11.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
